Fix a class cast exception when accessing 'pgtype.typlen' over wire protocol
(cherry picked from commit a1b08bce4a2a3d85036271bf72066ad2ba948dbb)

+++ sql/src/main/java/io/crate/protocols/postgres/types/PGType.java
@@ -50,8 +50,8 @@ public abstract class PGType {
         return oid;
     }
 
-    public int typeLen() {
-        return typeLen;
+    public short typeLen() {
+        return (short) typeLen;
     }
 
     public abstract int typArray();
